requirements.txt
# Core PyTorch with CUDA 12.x support for RTX 5080
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
--index-url https://download.pytorch.org/whl/cu124

# AI/ML Libraries with GPU acceleration
transformers>=4.44.0
accelerate>=0.33.0
diffusers>=0.30.0
sentence-transformers>=3.0.0
optimum>=1.21.0
xformers>=0.0.27  # Memory efficient attention for RTX 5080
safetensors>=0.4.0

# Computer Vision & Image Generation
opencv-python>=4.10.0
Pillow>=10.4.0
imageio>=2.34.0
imageio-ffmpeg>=0.5.1
scikit-image>=0.24.0

# Audio Processing & TTS
librosa>=0.10.2
soundfile>=0.12.1
pydub>=0.25.1
TTS>=0.22.0  # Coqui TTS for local voice synthesis
espeak-ng>=1.51  # TTS backend (optional)
phonemizer>=3.2.1
pyaudio>=0.2.14  # Audio I/O
resampy>=0.4.3  # Audio resampling

# Video Processing & FFmpeg
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# LLM & Text Processing
langchain>=0.2.16
langchain-community>=0.2.12
langchain-openai>=0.1.23
openai>=1.42.0
tiktoken>=0.7.0
beautifulsoup4>=4.12.3
markdownify>=0.12.1

# Data Science & Utils
numpy>=2.0.0
pandas>=2.2.2
scipy>=1.14.0
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.22.0

# Web Scraping & APIs
requests>=2.32.3
aiohttp>=3.10.0
feedparser>=6.0.11
python-dotenv>=1.0.1

# Database & Storage
chromadb>=0.5.5  # Vector database for content storage
sqlite3  # Built-in with Python
sqlalchemy>=2.0.32

# Scheduling & Automation
schedule>=1.2.2
celery>=5.4.0
redis>=5.0.8

# Configuration & Logging
pydantic>=2.8.2
PyYAML>=6.0.2
python-json-logger>=2.0.7
rich>=13.7.1  # Beautiful console output

# Development & Testing
pytest>=8.3.2
black>=24.8.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.11.1

# YouTube & Social Media APIs
google-api-python-client>=2.144.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

# Additional utilities
tqdm>=4.66.4  # Progress bars
psutil>=6.0.0  # System monitoring
packaging>=24.1
typing-extensions>=4.12.2
